Struct aws_sdk_mgn::client::fluent_builders::UpdateReplicationConfigurationTemplate

pub struct UpdateReplicationConfigurationTemplate<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateReplicationConfigurationTemplate<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateReplicationConfigurationTemplateOutput, SdkError<UpdateReplicationConfigurationTemplateError>> where
R::Policy: SmithyRetryPolicy<UpdateReplicationConfigurationTemplateInputOperationOutputAlias, UpdateReplicationConfigurationTemplateOutput, UpdateReplicationConfigurationTemplateError, UpdateReplicationConfigurationTemplateInputOperationRetryAlias>,

pub fn replication_configuration_template_id(
self,
inp: impl Into<String>
) -> Self

Update replication configuration template template ID request.

pub fn set_replication_configuration_template_id(
self,
input: Option<String>
) -> Self

pub fn arn(self, inp: impl Into<String>) -> Self

Update replication configuration template ARN request.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn staging_area_subnet_id(self, inp: impl Into<String>) -> Self

Update replication configuration template Staging Area subnet ID request.

pub fn set_staging_area_subnet_id(self, input: Option<String>) -> Self

pub fn associate_default_security_group(self, inp: bool) -> Self

Update replication configuration template associate default Application Migration Service Security group request.

pub fn set_associate_default_security_group(self, input: Option<bool>) -> Self

pub fn replication_servers_security_groups_i_ds(
self,
inp: impl Into<String>
) -> Self

Appends an item to replicationServersSecurityGroupsIDs.

To override the contents of this collection use set_replication_servers_security_groups_i_ds.

Update replication configuration template Replication Server Security groups IDs request.

pub fn set_replication_servers_security_groups_i_ds(
self,
input: Option<Vec<String>>
) -> Self

pub fn replication_server_instance_type(self, inp: impl Into<String>) -> Self

Update replication configuration template Replication Server instance type request.

pub fn set_replication_server_instance_type(self, input: Option<String>) -> Self

pub fn use_dedicated_replication_server(self, inp: bool) -> Self

Update replication configuration template use dedicated Replication Server request.

pub fn set_use_dedicated_replication_server(self, input: Option<bool>) -> Self

pub fn default_large_staging_disk_type(
self,
inp: ReplicationConfigurationDefaultLargeStagingDiskType
) -> Self

Update replication configuration template use default large Staging Disk type request.

pub fn set_default_large_staging_disk_type(
self,
input: Option<ReplicationConfigurationDefaultLargeStagingDiskType>
) -> Self

pub fn ebs_encryption(self, inp: ReplicationConfigurationEbsEncryption) -> Self

Update replication configuration template EBS encryption request.

pub fn set_ebs_encryption(
self,
input: Option<ReplicationConfigurationEbsEncryption>
) -> Self

pub fn ebs_encryption_key_arn(self, inp: impl Into<String>) -> Self

Update replication configuration template EBS encryption key ARN request.

pub fn set_ebs_encryption_key_arn(self, input: Option<String>) -> Self

pub fn bandwidth_throttling(self, inp: i64) -> Self

Update replication configuration template bandwidth throttling request.

pub fn set_bandwidth_throttling(self, input: Option<i64>) -> Self

pub fn data_plane_routing(
self,
inp: ReplicationConfigurationDataPlaneRouting
) -> Self

Update replication configuration template data plane routing request.

pub fn set_data_plane_routing(
self,
input: Option<ReplicationConfigurationDataPlaneRouting>
) -> Self

pub fn create_public_ip(self, inp: bool) -> Self

Update replication configuration template create Public IP request.

pub fn set_create_public_ip(self, input: Option<bool>) -> Self

pub fn staging_area_tags(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Adds a key-value pair to stagingAreaTags.

To override the contents of this collection use set_staging_area_tags.

Update replication configuration template Staging Area Tags request.

pub fn set_staging_area_tags(
self,
input: Option<HashMap<String, String>>
) -> Self