Module aws_sdk_mgn::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

throttling_exception

See ThrottlingException

uninitialized_account_exception

See UninitializedAccountException

validation_exception

See ValidationException

Structs

AccessDeniedException

Operating denied due to a file permission or access check error.

ChangeServerLifeCycleStateError

ConflictException

The request could not be completed due to a conflict with the current state of the target resource.

CreateReplicationConfigurationTemplateError

DeleteJobError

DeleteReplicationConfigurationTemplateError

DeleteSourceServerError

DescribeJobLogItemsError

DescribeJobsError

DescribeReplicationConfigurationTemplatesError

DescribeSourceServersError

DisconnectFromServiceError

FinalizeCutoverError

GetLaunchConfigurationError

GetReplicationConfigurationError

InitializeServiceError

InternalServerException

The server encountered an unexpected condition that prevented it from fulfilling the request.

ListTagsForResourceError

MarkAsArchivedError

ResourceNotFoundException

Resource not found exception.

RetryDataReplicationError

StartCutoverError

StartTestError

TagResourceError

TerminateTargetInstancesError

ThrottlingException

Reached throttling quota exception.

UninitializedAccountException

Unitialized account exception.

UntagResourceError

UpdateLaunchConfigurationError

UpdateReplicationConfigurationError

UpdateReplicationConfigurationTemplateError

ValidationException

Validate exception.

Enums

ChangeServerLifeCycleStateErrorKind

CreateReplicationConfigurationTemplateErrorKind

DeleteJobErrorKind

DeleteReplicationConfigurationTemplateErrorKind

DeleteSourceServerErrorKind

DescribeJobLogItemsErrorKind

DescribeJobsErrorKind

DescribeReplicationConfigurationTemplatesErrorKind

DescribeSourceServersErrorKind

DisconnectFromServiceErrorKind

FinalizeCutoverErrorKind

GetLaunchConfigurationErrorKind

GetReplicationConfigurationErrorKind

InitializeServiceErrorKind

ListTagsForResourceErrorKind

MarkAsArchivedErrorKind

RetryDataReplicationErrorKind

StartCutoverErrorKind

StartTestErrorKind

TagResourceErrorKind

TerminateTargetInstancesErrorKind

UntagResourceErrorKind

UpdateLaunchConfigurationErrorKind

UpdateReplicationConfigurationErrorKind

UpdateReplicationConfigurationTemplateErrorKind